grafo orientato e pesato

+++ model/Arco.py
@@ -2,7 +2,8 @@ from dataclasses import dataclass
 
 
 @dataclass
+
 class Arco:
-    ordine1:int
-    ordine2:int
-    peso:float
+    ordine1: int
+    ordine2: int
+    peso: int
